Change pin number to u8

+++ hermes-five/src/devices/input/analog.rs
@@ -21,7 +21,7 @@ pub struct AnalogInput {
     // ########################################
     // # Basics
     /// The pin (id) of the [`Board`] used to read the analog value.
-    pin: u16,
+    pin: u8,
     /// The current AnalogInput state.
     #[cfg_attr(feature = "serde", serde(with = "crate::devices::arc_rwlock_serde"))]
     state: Arc<RwLock<u16>>,
@@ -71,7 +71,7 @@ impl AnalogInput {
     }
 
     /// Returns the pin (id) used by the device.
-    pub fn get_pin(&self) -> u16 {
+    pub fn get_pin(&self) -> u8 {
         self.pin
     }
