+++ hermes-five/src/devices/output/pwm.rs
@@ -19,7 +19,7 @@ pub struct PwmOutput {
     // ########################################
     // # Basics
     /// The pin (id) of the [`Board`] used to control the output value.
-    pin: u16,
+    pin: u8,
     /// The current output state.
     #[cfg_attr(feature = "serde", serde(with = "crate::devices::arc_rwlock_serde"))]
     state: Arc<RwLock<u16>>,
@@ -88,7 +88,7 @@ impl PwmOutput {
     // Setters and Getters.
 
     /// Returns the pin (id) used by the device.
-    pub fn get_pin(&self) -> u16 {
+    pub fn get_pin(&self) -> u8 {
         self.pin
     }
