+++ hermes-five/src/io/protocol.rs
@@ -44,19 +44,19 @@ pub trait IoProtocol: DynClone + Send + Sync + Debug + Display {
     /// Sets the `mode` of the specified `pin`.
     ///
     /// <https://github.com/firmata/protocol/blob/master/protocol.md#data-message-expansion>
-    fn set_pin_mode(&mut self, pin: u16, mode: PinModeId) -> Result<(), Error>;
+    fn set_pin_mode(&mut self, pin: u8, mode: PinModeId) -> Result<(), Error>;
 
     /// Writes `level` to the digital `pin`.
     ///
     /// Send an DIGITAL_MESSAGE (0x90 - set digital value).
     /// <https://github.com/firmata/protocol/blob/master/protocol.md#message-types>
-    fn digital_write(&mut self, pin: u16, level: bool) -> Result<(), Error>;
+    fn digital_write(&mut self, pin: u8, level: bool) -> Result<(), Error>;
 
     /// Writes `level` to the analog `pin`.
     ///
     /// Send an ANALOG_MESSAGE (0xE0 - set analog value).
     /// <https://github.com/firmata/protocol/blob/master/protocol.md#message-types>
-    fn analog_write(&mut self, pin: u16, level: u16) -> Result<(), Error>;
+    fn analog_write(&mut self, pin: u8, level: u16) -> Result<(), Error>;
 
     ///  Sets the analog reporting `state` of the specified analog `pin`.
     ///
@@ -73,7 +73,7 @@ pub trait IoProtocol: DynClone + Send + Sync + Debug + Display {
     ///
     /// This will activate the reporting of all pins in port (hence the pin will send us its value periodically)
     /// <https://github.com/firmata/protocol/blob/master/protocol.md>
-    fn report_digital(&mut self, pin: u16, state: bool) -> Result<(), Error>;
+    fn report_digital(&mut self, pin: u8, state: bool) -> Result<(), Error>;
 
     /// Set the sampling `interval` (in ms).
     ///
@@ -89,7 +89,7 @@ pub trait IoProtocol: DynClone + Send + Sync + Debug + Display {
 
     /// Sends a SERVO_CONFIG command (0x70 - configure servo)
     /// <https://github.com/firmata/protocol/blob/master/servos.md>
-    fn servo_config(&mut self, pin: u16, pwm_range: Range<u16>) -> Result<(), Error>;
+    fn servo_config(&mut self, pin: u8, pwm_range: Range<u16>) -> Result<(), Error>;
 
     // ########################################
     // I2C
@@ -98,9 +98,9 @@ pub trait IoProtocol: DynClone + Send + Sync + Debug + Display {
     /// register is written to and the data in that register can be read.
     fn i2c_config(&mut self, delay: u16) -> Result<(), Error>;
     /// Reads `size` bytes from I2C device at the specified `address`.
-    fn i2c_read(&mut self, address: i32, size: i32) -> Result<(), Error>;
+    fn i2c_read(&mut self, address: u8, size: u8) -> Result<(), Error>;
     /// Writes `data` to the I2C device at the specified `address`.
-    fn i2c_write(&mut self, address: i32, data: &[u8]) -> Result<(), Error>;
+    fn i2c_write(&mut self, address: u8, data: &[u8]) -> Result<(), Error>;
 }
 
 #[cfg(not(tarpaulin_include))]
